Moving merkle_sign, merkle_gen_root to merkle.c
